backend/requirements.txt
fastapi[all]
pydantic
sqlalchemy
python-dotenv